optee: don't warn about missing OP-TEE header
OP-TEE header is checked once in PBL, saved into scratch area after
verification and then checked again in barebox proper.

The check in PBL fails silently, but the check in barebox proper that
should always follow, because the header isn't written to the scratch
area is printed with error log level.

Printing an error in either case is wrong though as using a raw OP-TEE
binary without header is a valid use case and the OP-TEE header may
also be missing when barebox is chainloaded from a running barebox.

Therefore reduce the message to debug log level.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Reviewed-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20240228180319.2739412-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>