FIT: support finding compatible configuration by FDT compatible
So far, we only supported finding compatible configurations that have
a compatible property inside the configuration's device tree node.

According to spec, this is optional however, and e.g. Yocto's
kernel-fitimage.bbclass don't generate it.

Instead, the bootloader is expected to lookup the compatible inside the
referenced FDT. With fdt_machine_is_compatible, this is much less of a
performance hit than with of_machine_is_compatible, so let's implement
support for this.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240301130445.171385-5-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>