efi-stdio: fix wait_for_event argument
EFI v2.10 documents the following about EFI events:

  | // EFI_EVENT: Handle to an event structure. Type VOID *
  |
  | typedef
  | EFI_STATUS
  | (EFIAPI *EFI_WAIT_FOR_EVENT) (
  | 	IN UINTN NumberOfEvents,
  | 	IN EFI_EVENT *Event,
  | 	OUT UINTN *Index
  | );
  |
  | typedef struct _EFI_SIMPLE_TEXT_INPUT_PROTOCOL {
  |     EFI_INPUT_RESET Reset;
  |     EFI_INPUT_READ_KEY ReadKeyStroke;
  |     EFI_EVENT WaitForKey;
  | } EFI_SIMPLE_TEXT_INPUT_PROTOCOL;

To sum up, wait_for_event takes the number of events to wait for and a
pointer to an EFI_EVENT array of that size. Because we define efi_event
as void *, it went unnoticed that we passed a plain pointer instead of
a pointer to a pointer like the API expects.

With the using of an opaque type in the follow-up commit, this will
trigger a warning, so we fix this here in anticipation.

I am not sure how this went unnoticed so far, but the efi-stdio
console driver behaves as one would expect in Qemu.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240304190038.3486881-73-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>