efi: payload: early-mem: simplify error message reporting
Adding a size to the error message means that we can just use a single
error message for either error case with no loss of information.

While at it, we turn the error string into a code: This file can be
built for PBL later on and we won't have efi_strerror there.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240304190038.3486881-85-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>