console: scale number of log messages with available RAM
The default value of 1000 messages is too small with
CONFIG_COMPILE_LOGLEVEL=7, because the many driver debug messages
will evict any earlier error or warning messages that have been
in the log.

Assuming a log record size of 128 bytes, 1000 messages translates to
128kB, which has quite a bit of breathing room. Let's instead scale the
maximum with the available memory: one message per 32KiB of malloc area.

This gives us:

  -  64M total memory ->  ~32M malloc area ->  1K messages (128K bytes)
  - 256M total memory -> ~128M malloc area ->  4K messages (512K bytes)
  -   1G total memory -> ~512M malloc area -> 16K messages (  2M bytes)

Users for which this is too excessive can still configure this via the
environment.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240306184519.626594-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>