optee: build common code depending on CONFIG_HAVE_OPTEE
There's a stub for optee_set_membase(), but it's only defined when
CONFIG_HAVE_OPTEE is disabled. The out-of-line version is only linked
when there's either PBL_OPTEE or BOOTM_OPTEE, both of which select
HAVE_OPTEE.

HAVE_OPTEE can also be selected by OPTEE, which is the option for
the driver communicating with OP-TEE. This leads to a linker error
with the combination of CONFIG_OPTEE=y, CONFIG_PBL_OPTEE=n and
CONFIG_BOOTM_OPTEE=n.

Fix this by using CONFIG_HAVE_OPTEE both in header and Makefile.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240306193936.712501-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>