common: console_common: handle NULL dev gracefully in dev_printf
dev_name(NULL) returns NULL, which snprintf can handle, but dev->driver
used beforehand would derference the NULL pointer without checking.

Fix this by checking that dev is !NULL before dereferencing.
This still leads to an ugly <NULL>: prefix, when a NULL dev is used, so
it can be fixed, but without the crash, which might be difficult to
pinpoint if the print happens early.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240313092633.3339320-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>