pmdomain: imx: gpc2: ignore missing regulators
The GPV2 driver supports controlling voltage rails associated with
domains via adding a power-supply binding to the power domain's
device tree node. This is only done for the i.MX8MQ's GPU, VPU and MIPI
domains. All boards supported upstream that do this reference a
regulator of the pfuze100 as power supply.

While we have a pfuze100 driver, it only does one-time init and doesn't
implement a regulator API. This means these boards will fail with probe
deferrals:

  imx-pgc imx-pgc-domain5: error Requested probe deferral: Failed to get domain's regulator
  imx-pgc imx-pgc-domain6: error Requested probe deferral: Failed to get domain's regulator
  ERROR: imx-pgc-domain6: probe permanently deferred (Failed to get domain's regulator)
  ERROR: imx-pgc-domain5: probe permanently deferred (Failed to get domain's regulator)

As we don't use MIPI, GPU or VPU in barebox and we want to move boards
over to deep probe (where every EPROBE_DEFER is permanent), let's stop
propagating EPROBE_DEFER and just write an info message and continue
probing as usual and delay the error message until the regulator is
actually used.

Fixes: eb487a8d39e2 ("soc: imx: gpcv2: extend for i.MX8M Mini/Nano/Plus support")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240313144136.2157496-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>