include: kconfig.h: move __is_defined definition into new header
While we run the preprocessor over device trees before compiling them,
we don't define the Kconfig symbols and including <generated/autoconf.h>
would fail.

The file defines __is_defined however, which can be useful for other
macros that are either defined to 1 or undefined like the macros we
define for device tree fragments to indicate which file they are being
appended to. To make it possible to use __is_defined from fragments,
move the kconfig-independent part out of <linux/kconfig.h> into
<linux/is_defined.h>.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240403072332.2694292-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>