mci: improve packing of struct mci
rca will be at least 4-byte aligned due to being after a 4-byte member.
It's only 2 byte long though and followed by another 4 byte member,
wasting 2 bytes. Fill those up with the two 1-byte bool members we have.

This includes a change of type for high_capacity: This is ok, because
only 0 or 1 are ever saved into it.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240403090353.2976380-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>