mci: remove FIXME in rca description
The FIXME was added in the commit introducing the file many years ago.
The RCA is a bit special in that it's unused for SPI, reported by the
card for MMC and set by the host for SD. There are comments describing
this in the MCI core, so just explain the abbreviation in the
documentation comment.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240403090353.2976380-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>